:recycle:Refactor : 글작성 2개 페이지 1개로 수정 (#126)
